libs/http-ext/src/lib/matchers/match-origin/get-origin.spec.ts
import { getOrigin } from './get-origin';

describe('getOrigin', () => {
  it('should extract origin from URL', () => {
    expect(getOrigin('https://jscutlery.github.io')).toEqual(
      'https://jscutlery.github.io'
    );
    expect(getOrigin('https://jscutlery.github.io/test')).toEqual(
      'https://jscutlery.github.io'
    );
    expect(getOrigin('https://jscutlery.github.io:443/test')).toEqual(
      'https://jscutlery.github.io:443'
    );
  });

  it('should fail if invalid URL', () => {
    expect(() => getOrigin('jscutlery.github.io')).toThrow();
    expect(() => getOrigin('jscutlery.github.io:443')).toThrow();
    expect(() => getOrigin('/test')).toThrow();
  });
});
